refactor: scope exclude normalization to source, hoist platform test helpers
_worktree_local_exclude no longer blanket-rewrites every pattern's separators;
only the glob-derived rel can carry os.sep, so it's normalized at the source
via rel.as_posix(). skill_tree/config_path/seed_files are already POSIX config
strings, so the fix stays precise and never touches a pattern that could
pathologically hold a legitimate backslash on POSIX.

Hoist the host-shell verify/lifecycle stub helpers (_OK, _RUN, _file_exists_cmd,
_touch_run, _exists_run, _seeded_then_touch) into conftest so the sys.platform
branching lives in one spot instead of being re-derived per test file; the two
engine test modules import them and drop their now-unused sys import.
