+++ tests/conftest.py
@@ -45,6 +45,51 @@ def write_script_launcher(directory: Path, name: str, body: str) -> Path:
     return launcher
 
 
+# ---- host-shell verify/lifecycle stub commands (single platform-detection spot) ----
+# The engine runs verify/plugin-lifecycle commands via the host shell (`sh -c` on
+# POSIX, `cmd /c` on Windows), so tests that assert on that machinery need commands
+# both shells honor. These build them per-OS in one place instead of each test file
+# re-deriving the win32 branch.
+
+_OK = "exit 0"  # cross-platform always-success verb (both `cmd /c` and `sh -c` honor it)
+_RUN = "%BMAD_AUTO_RUN_DIR%" if sys.platform == "win32" else "$BMAD_AUTO_RUN_DIR"
+
+
+def _file_exists_cmd(path) -> str:
+    """Shell verify command (run via shell=True) exiting 0 iff `path` exists, on the
+    host's shell — `test -f` (POSIX) / `if exist` (Windows cmd) — so the verify-gate
+    tests drive the real machinery on either OS, not a POSIX-only `test` that cmd
+    rejects with "'test' is not recognized"."""
+    if sys.platform == "win32":
+        return f'if exist "{path}\\NUL" (exit 1) else if exist "{path}" (exit 0) else (exit 1)'
+    return f'test -f "{path}"'
+
+
+def _touch_run(marker: str) -> str:
+    if sys.platform == "win32":
+        return f'type nul > "{_RUN}\\{marker}"'
+    return f'touch "{_RUN}/{marker}"'
+
+
+def _exists_run(marker: str) -> str:
+    if sys.platform == "win32":
+        return (
+            f'if exist "{_RUN}\\{marker}\\NUL" (exit 1) '
+            f'else if exist "{_RUN}\\{marker}" (exit 0) else (exit 1)'
+        )
+    return f'test -f "{_RUN}/{marker}"'
+
+
+def _seeded_then_touch(rel: str, marker: str) -> str:
+    if sys.platform == "win32":
+        norm_rel = rel.replace("/", "\\")
+        return (
+            f'if exist "{norm_rel}\\NUL" (exit 1) '
+            f'else if exist "{norm_rel}" (type nul > "{_RUN}\\{marker}") else (exit 1)'
+        )
+    return f'test -f "{rel}" && touch "{_RUN}/{marker}"'
+
+
 SPRINT_TEMPLATE = {
     "generated": "01-06-2026 10:00",
     "last_updated": "01-06-2026 10:00",
